+++ boussole/conf/base_backend.py
@@ -39,6 +39,7 @@ class SettingsBackendBase(SettingsPostProcessor):
     """
     _default_filename = 'settings.txt' #: Default filename
     _kind_name = 'txt' #: Backend format name
+    _file_extension = 'txt' #: Default filename extension
 
     def __init__(self, basedir=None):
         self.basedir = basedir or ''
@@ -117,17 +118,33 @@ class SettingsBackendBase(SettingsPostProcessor):
 
     def parse(self, filepath, content):
         """
-        Parse opened settings content
+        Load and parse opened settings content.
 
         Base method do nothing because parsing is dependent from backend.
 
         Args:
-            filepath (str): Settings object, depends from backend
+            filepath (str): Settings file location.
             content (str): Settings content from opened file, depends from
                 backend.
 
         Returns:
-            dict: Dictionnary containing parsed setting elements.
+            dict: Dictionnary containing parsed setting options.
+
+        """
+        return {}
+
+    def dump(self, content, filepath):
+        """
+        Dump settings content to filepath.
+
+        Base method do nothing because dumping is dependent from backend.
+
+        Args:
+            content (str): Settings content.
+            filepath (str): Settings file location.
+
+        Returns:
+            dict: Dictionnary containing parsed setting options.
 
         """
         return {}
@@ -158,7 +175,7 @@ class SettingsBackendBase(SettingsPostProcessor):
             filepath (str): Filepath to the settings file.
 
         Returns:
-            boussole.conf.model.Settings: Settings object with loaded elements.
+            boussole.conf.model.Settings: Settings object with loaded options.
 
         """
         self.projectdir, filename = self.parse_filepath(filepath)