+++ boussole/conf/json_backend.py
@@ -15,6 +15,18 @@ class SettingsBackendJson(SettingsBackendBase):
     """
     _default_filename = 'settings.json' #: Default filename
     _kind_name = 'json' #: Backend format name
+    _file_extension = 'json' #: Default filename extension
+
+    def dump(self, content, filepath, indent=4):
+        """
+        Dump settings content to filepath.
+
+        Args:
+            content (str): Settings content.
+            filepath (str): Settings file location.
+        """
+        with open(filepath, 'w') as fp:
+            json.dump(content, fp, indent=indent)
 
     def parse(self, filepath, content):
         """