+++ tests/102_projectstarter/004_init.py
@@ -20,7 +20,7 @@ def test_success(projectstarter, temp_builds_dir, name, ext, module):
     basedir = temp_builds_dir.join(tmp_dirname).strpath
     os.makedirs(basedir)
 
-    results = projectstarter(backend=name).init(*(
+    results = projectstarter(backend_name=name).init(*(
         '.',
         settings_filename,
         'scss',
@@ -62,7 +62,7 @@ def test_error(projectstarter, temp_builds_dir, name, ext, module):
     os.makedirs(basedir)
 
     with pytest.raises(SettingsInvalidError):
-        results = projectstarter(backend=name).init(*(
+        results = projectstarter(backend_name=name).init(*(
             '.',
             settings_filename,
             'css',